btrfs-progs: image: move defintions from main.c to header
